Add modules folder with files
