+++ index.html
@@ -5,7 +5,7 @@
     <meta http-equiv="X-UA-Compatible" content="IE=edge">
     <meta name="viewport" content="width=device-width, initial-scale=1.0">
     <title>Awesome Books</title>
-    <link rel="stylesheet" href="style.css">
+    <link rel="stylesheet" href="index.css">
     <link rel="stylesheet" href="https://unicons.iconscout.com/release/v4.0.0/css/line.css">
 </head>
 <body>
@@ -58,6 +58,6 @@
         <span>&copy</span>
         <p>2022 Sane Myburg</p>    
     </footer>
-    <script src="script.js"></script>
+    <script type="module" src="./index.js"></script>
 </body>
 </html>
