+++ modules/storage.js
@@ -1,0 +1,28 @@
+// Storing Handling
+export default class Store {
+    static getBooks() {
+      let books;
+      if (localStorage.getItem('books') === null) {
+        books = [];
+      } else {
+        books = JSON.parse(localStorage.getItem('books'));
+      }
+      return books;
+    }
+  
+    static addBook(book) {
+      const books = Store.getBooks();
+      books.push(book);
+      localStorage.setItem('books', JSON.stringify(books));
+    }
+  
+    static removeBook(author) {
+      const books = Store.getBooks();
+      books.forEach((book, index) => {
+        if (book.author === author) {
+          books.splice(index, 1);
+        }
+      });
+      localStorage.setItem('books', JSON.stringify(books));
+    }
+  }
